+++ ruby/lib/gitlab/git.rb
@@ -8,6 +8,7 @@ require 'active_support/core_ext/object/blank'
 require 'active_support/core_ext/numeric/bytes'
 require 'active_support/core_ext/numeric/time'
 require 'active_support/core_ext/module/delegation'
+require 'active_support/core_ext/hash/transform_values'
 require 'active_support/core_ext/enumerable'
 
 # We split our mock implementation of Gitlab::GitalyClient into a separate file